requirements.txt
Django==5.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
Pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
